mcp-qrcode/pyproject.toml
[project]
name = "mcp-qrcode"
version = "1.0.0"
description = "Generate and decode QR codes, vCards, WiFi configs, styled QR codes"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "qrcode[pil]>=7.4",
    "Pillow>=10.0.0",
    "pyzbar>=0.1.9",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
mcp-qrcode = "src.server:main"
